feat: remove duplicated logic
