+++ src/index.ts
@@ -6,7 +6,7 @@ import type { UnistylesThemes, UnistylesBreakpoints } from './global'
 import { ScreenOrientation, AndroidContentSizeCategory, IOSContentSizeCategory } from './common'
 import { useStyles } from './useStyles'
 import { createStyleSheet } from './createStyleSheet'
-import { UnistylesProvider } from './context'
+import { UnistylesProvider } from './UnistylesProvider'
 /**
  * Utility to interact with the Unistyles
  * (should be called only once)
